refactor: Moved Axios Calls to App.js to Pass Results  Between Components Easier.
