+++ frontend/src/App.js
@@ -1,5 +1,9 @@
 // General Imports
 import { Routes, Route } from "react-router-dom";
+import { useState, useEffect } from "react";
+import axios from "axios";
+import { KEY } from "./local_key";
+import { ID } from "./local_id";
 import "./App.css";
 
 // Pages Imports
@@ -15,6 +19,44 @@ import Footer from "./components/Footer/Footer";
 import PrivateRoute from "./utils/PrivateRoute";
 
 function App() {
+
+  const [accessToken, setAccessToken] = useState("");
+  const [results, setResults] = useState({})
+
+  const getAuth = async () => {
+    try {
+      let response = await axios.post(`https://api.petfinder.com/v2/oauth2/token`, `grant_type=client_credentials&client_id=${ID}&client_secret=${KEY}`, {
+        headers: {
+          'Content-Type': `application/x-www-form-urlencoded`
+        }
+      });
+      setAccessToken(response.data.access_token)
+      console.log(response.data)
+    }
+    catch(error){
+      console.log(error.message)
+    }
+  };
+
+  const getResults = async (props) => {
+    try {
+      let response = await axios.get(`https://api.petfinder.com/v2/animals?type=${props.animalType}&location=${props.zipCode}&distance=50`, {
+        headers: {
+          'Authorization': `Bearer ${accessToken}`
+        }
+      });
+      setResults(response.data)
+      console.log(response.data)
+    }
+    catch(error){
+      console.log(error.message)
+    }
+  }
+
+  useEffect(() => {
+    getAuth();
+  }, []);
+
   return (
     <div>
       <Navbar />
@@ -23,7 +65,7 @@ function App() {
           path="/"
           element={
             <PrivateRoute>
-              <HomePage />
+              <HomePage getResults={getResults}/>
             </PrivateRoute>
           }
         />
